Split CPU and OpenCL scan and scan_by_key files
