+++ src/backend/cuda/scan_by_key_add.cu
@@ -11,6 +11,6 @@
 
 namespace cuda
 {
-    INSTANTIATE_SCAN_OP(af_add_t)
+    INSTANTIATE_SCAN_BY_KEY_ALL_OP(af_add_t)
 }
