+++ src/backend/cuda/scan_by_key_mul.cu
@@ -11,6 +11,6 @@
 
 namespace cuda
 {
-    INSTANTIATE_SCAN_OP(af_mul_t)
+    INSTANTIATE_SCAN_BY_KEY_ALL_OP(af_mul_t)
 }
