+++ src/backend/opencl/scan_add.cpp
@@ -7,10 +7,10 @@
  * http://arrayfire.com/licenses/BSD-3-Clause
  ********************************************************/
 
-#include <scan_by_key_impl.hpp>
+#include <scan_impl.hpp>
 
-namespace cuda
+namespace opencl
 {
-    INSTANTIATE_SCAN_BY_KEY_ALL_OP(af_max_t)
+    INSTANTIATE_SCAN(af_notzero_t, char, uint)
+    INSTANTIATE_SCAN_ALL(af_add_t)
 }
-
